+++ sensor_msgs/include/sensor_msgs/point_cloud_iterator.h
@@ -45,49 +45,6 @@
  * \author Vincent Rabaud
  */
 
-namespace
-{
-/** Private function that adds a PointField to the "fields" member of a PointCloud2
- * @param cloud_msg the PointCloud2 to add a field to
- * @param name the name of the field
- * @param count the number of elements in the PointField
- * @param datatype the datatype of the elements
- * @param offset the offset of that element
- * @return the offset of the next PointField that will be added to the PointCLoud2
- */
-inline int addPointField(sensor_msgs::PointCloud2* cloud_msg, const std::string &name, int count, int datatype,
-    int offset)
-{
-  sensor_msgs::PointField point_field;
-  point_field.name = name;
-  point_field.count = count;
-  point_field.datatype = datatype;
-  point_field.offset = offset;
-  cloud_msg->fields.push_back(point_field);
-
-  // Update the offset
-  int point_field_type;
-  if (datatype == sensor_msgs::PointField::INT8)
-    point_field_type = sizeof(int8_t);
-  else if (datatype == sensor_msgs::PointField::UINT8)
-    point_field_type = sizeof(uint8_t);
-  else if (datatype == sensor_msgs::PointField::INT16)
-    point_field_type = sizeof(int16_t);
-  else if (datatype == sensor_msgs::PointField::UINT16)
-    point_field_type = sizeof(uint16_t);
-  else if (datatype == sensor_msgs::PointField::INT32)
-    point_field_type = sizeof(int32_t);
-  else if (datatype == sensor_msgs::PointField::UINT32)
-    point_field_type = sizeof(uint32_t);
-  else if (datatype == sensor_msgs::PointField::FLOAT32)
-    point_field_type = sizeof(float);
-  else
-    point_field_type = sizeof(double);
-
-  return offset + point_field.count * point_field_type / sizeof(char);
-}
-}
-
 namespace sensor_msgs
 {
 /** Function setting some fields in a PointCloud and adjusting the internals of the PointCloud2
@@ -103,23 +60,7 @@ namespace sensor_msgs
  *          number of elements in the field, the datatype of the elements in the field
  * @return a reference to the original PointCloud2 but modified
  */
-inline void setPointCloud2Fields(sensor_msgs::PointCloud2* cloud_msg, int n_fields, ...)
-{
-  cloud_msg->fields.clear();
-  cloud_msg->fields.reserve(n_fields);
-  va_list vl;
-  va_start(vl, n_fields);
-  int offset = 0;
-  for (int i = 0; i < n_fields; ++i)
-    // Create the corresponding PointField
-    offset = addPointField(cloud_msg, std::string(va_arg(vl, char *)), va_arg(vl, int), va_arg(vl, int), offset);
-    va_end(vl);
-
-  // Resize the point cloud accordingly
-  cloud_msg->point_step = offset;
-  cloud_msg->row_step = cloud_msg->width * cloud_msg->point_step;
-  cloud_msg->data.resize(cloud_msg->height * cloud_msg->row_step);
-}
+void setPointCloud2Fields(sensor_msgs::PointCloud2* cloud_msg, int n_fields, ...);
 
 /** Function setting some fields in a PointCloud and adjusting the internals of the PointCloud2
  * @param cloud_msg
@@ -128,44 +69,12 @@ inline void setPointCloud2Fields(sensor_msgs::PointCloud2* cloud_msg, int n_fiel
  *            stacked in a float), "rgba" (4 uchar stacked in a float)
  * @return a reference to the original PointCloud2 but modified
  */
-inline void setPointCloud2FieldsByString(sensor_msgs::PointCloud2* cloud_msg, int n_fields, ...)
-{
-  cloud_msg->fields.clear();
-  cloud_msg->fields.reserve(n_fields);
-  va_list vl;
-  va_start(vl, n_fields);
-  int offset = 0;
-  for (int i = 0; i < n_fields; ++i)
-  {
-    // Create the corresponding PointFields
-    std::string
-    field_name = std::string(va_arg(vl, char *));
-    if (field_name == "xyz")
-    {
-      sensor_msgs::PointField point_field;
-      // Do x, y and z
-      offset = addPointField(cloud_msg, "x", 1, sensor_msgs::PointField::FLOAT32, offset);
-      offset = addPointField(cloud_msg, "y", 1, sensor_msgs::PointField::FLOAT32, offset);
-      offset = addPointField(cloud_msg, "z", 1, sensor_msgs::PointField::FLOAT32, offset);
-    }
-    else if ((field_name == "rgb") || (field_name == "rgba"))
-    {
-      offset = addPointField(cloud_msg, field_name, 1, sensor_msgs::PointField::FLOAT32, offset);
-    }
-    else
-      throw std::runtime_error("Field " + field_name + " does not exist");
-  }
-  va_end(vl);
-
-  // Resize the point cloud accordingly
-  cloud_msg->point_step = offset;
-  cloud_msg->row_step = cloud_msg->width * cloud_msg->point_step;
-  cloud_msg->data.resize(cloud_msg->height * cloud_msg->row_step);
+void setPointCloud2FieldsByString(sensor_msgs::PointCloud2* cloud_msg, int n_fields, ...);
 }
 
 namespace
 {
-/** Private base class that of PointCloud2Iterator and PointCloud2ConstIterator
+/** Private base class for PointCloud2Iterator and PointCloud2ConstIterator
  * T is the type of the value to be retrieved
  * U is the type of the raw data in PointCloud2 (only uchar and const uchar are supported)
  */
@@ -177,113 +86,53 @@ public:
    * @param cloud_msg The PointCloud2 to iterate upon
    * @param field_name The field to iterate upon
    */
-  void initialize(sensor_msgs::PointCloud2 &cloud_msg, const std::string &field_name)
-  {
-    int offset = set_fields(cloud_msg, field_name);
-
-    data_char_ = &(cloud_msg.data.front()) + offset;
-    data_ = reinterpret_cast<T*>(data_char_);
-    data_end_ = reinterpret_cast<T*>(&(cloud_msg.data.back()) + 1 + offset);
-  }
+  void initialize(sensor_msgs::PointCloud2 &cloud_msg, const std::string &field_name);
 
   /** Const version of the above
    * @param cloud_msg The PointCloud2 to iterate upon
    * @param field_name The field to iterate upon
    */
-  void initialize(const sensor_msgs::PointCloud2 &cloud_msg, const std::string &field_name)
-  {
-    int offset = set_fields(cloud_msg, field_name);
-
-    data_char_ = &(cloud_msg.data.front()) + offset;
-    data_ = reinterpret_cast<T*>(data_char_);
-    data_end_ = reinterpret_cast<T*>(&(cloud_msg.data.back()) + 1 + offset);
-  }
+  void initialize(const sensor_msgs::PointCloud2 &cloud_msg, const std::string &field_name);
 
   /** Access the i th element starting at the current pointer (useful when a field has several elements of the same
    * type)
    * @param i
    * @return a reference to the i^th value from the current position
    */
-  T& operator [](size_t i) const
-  {
-    if (is_bigendian_)
-      return *(data_ - i);
-    else
-      return *(data_ + i);
-  }
+  T& operator [](size_t i) const;
 
   /** Dereference the iterator. Equivalent to accessing it through [0]
    * @return the value to which the iterator is pointing
    */
-  T& operator *() const
-  {
-    return *data_;
-  }
+  T& operator *() const;
 
   /** Increase the iterator to the next element
    * @return a reference to the updated iterator
    */
-  PointCloud2IteratorBase<T, U>& operator ++()
-  {
-    data_char_ += point_step_;
-    data_ = reinterpret_cast<T*>(data_char_);
-    return *this;
-  }
+  PointCloud2IteratorBase<T, U>& operator ++();
 
-    /** Increase the iterator of a ceratin amount
+  /** Increase the iterator of a ceratin amount
    * @return a reference to the updated iterator
    */
-  PointCloud2IteratorBase<T, U>& operator +(int i)
-  {
-    data_char_ += i*point_step_;
-    data_ = reinterpret_cast<T*>(data_char_);
-    return *this;
-  }
+  PointCloud2IteratorBase<T, U>& operator +(int i);
 
   /** Compare to another iterator
    * @return whether the current iterator points to a different address than the other one
    */
-  bool operator !=(const PointCloud2IteratorBase<T, U>& iter) const
-  {
-    return iter.data_ != data_;
-  }
+  bool operator !=(const PointCloud2IteratorBase<T, U>& iter) const;
 
   /** Return the end iterator
    * @return the end iterator (useful when performing normal iterator processing with ++)
    */
-  PointCloud2IteratorBase<T, U> end() const
-  {
-    PointCloud2IteratorBase<T, U> res;
-    res.data_ = data_end_;
-    return res;
-  }
+  PointCloud2IteratorBase<T, U> end() const;
+
 private:
   /** Common code to set the fiels of the point cloud
    * @param cloud_msg the PointCloud2 to modify
    * @param field_name the name of the field to iterate upon
    * @return the offset at which the field is found
    */
-  int set_fields(const sensor_msgs::PointCloud2 &cloud_msg, const std::string &field_name)
-  {
-    is_bigendian_ = cloud_msg.is_bigendian;
-    point_step_ = cloud_msg.point_step;
-    // make sure the channel is valid
-    std::vector<sensor_msgs::PointField>::const_iterator field_iter = cloud_msg.fields.begin(), field_end =
-        cloud_msg.fields.end();
-    while ((field_iter != field_end) && (field_iter->name != field_name))
-      ++field_iter;
-
-    if (field_iter == field_end)
-      throw std::runtime_error("Field " + field_name + " does not exist");
-
-    // continue filling the info
-    field_count_ = field_iter->count;
-
-    if (is_bigendian_)
-      return point_step_ - field_iter->offset;
-    else
-      return field_iter->offset;
-  }
+  int set_fields(const sensor_msgs::PointCloud2 &cloud_msg, const std::string &field_name);
 
   /** The "point_step" of the original cloud */
   int point_step_;
@@ -300,6 +149,8 @@ private:
 };
 }
 
+namespace sensor_msgs
+{
 /** Class that can iterate over a PointCloud2
  * T type of the element being iterated upon
  * E.g, you create your PointClou2 message as follows:
@@ -321,10 +172,7 @@ public:
    * @param cloud_msg The PointCloud2 to iterate upon
    * @param field_name The field to iterate upon
    */
-  PointCloud2Iterator(sensor_msgs::PointCloud2 &cloud_msg, const std::string &field_name)
-  {
-    PointCloud2IteratorBase<T, unsigned char>::initialize(cloud_msg, field_name);
-  }
+  PointCloud2Iterator(sensor_msgs::PointCloud2 &cloud_msg, const std::string &field_name);
 };
 
 /** Same as a PointCloud2Iterator but for const data
@@ -336,11 +184,10 @@ public:
    * @param cloud_msg The PointCloud2 to iterate upon
    * @param field_name The field to iterate upon
    */
-  PointCloud2ConstIterator(const sensor_msgs::PointCloud2 &cloud_msg, const std::string &field_name)
-  {
-    PointCloud2IteratorBase<const T, const unsigned char>::initialize(cloud_msg, field_name);
-  }
+  PointCloud2ConstIterator(const sensor_msgs::PointCloud2 &cloud_msg, const std::string &field_name);
 };
-
 }
+
+#include <sensor_msgs/impl/point_cloud_iterator.h>
+
 #endif// SENSOR_MSGS_POINT_CLOUD_ITERATOR_H
